backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers>=4.35.2
torch>=2.6.0
tokenizers>=0.15.0
gtts==2.4.0
moviepy==1.0.3
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
pillow>=10.1.0
numpy>=1.24.3